requirements.txt
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.3
Pillow==7.0.0
pyaml==20.3.1
scikit-learn==0.22.1
scipy==1.4.1
torch==1.4.0
torchvision==0.5.0
tqdm==4.44.1
visdom==0.1.8.9
